requirements.txt
# --- cœur RAG ---
langchain>=0.2
langchain-community>=0.2
langchain-openai>=0.1
langchain-chroma>=0.1
chromadb>=0.4
openai>=1.25
tiktoken>=0.6

# --- tooling ---
gitpython>=3.1
python-dotenv>=1.0
pydantic>=2.7
structlog>=24.1

# --- frontend ---
streamlit>=1.35

# --- dev ---
pytest>=8.2
black>=24.4
ruff>=0.4
isort>=5.13
mypy>=1.10
pre-commit>=3.7